Multimodal/AwsCloudDialogueAnalyzer/requirements.txt
boto3
google-auth
google-auth-oauthlib
google-auth-httplib2
firebase-admin
google-api-python-client